emule: support Blackhole's 64 circular buffers
Ziveli generic_op softmax/attention kernels (and any wide-CB program, e.g. a
36-source concat) allocate circular buffers at ids up to ~53. Blackhole has 64
CBs (Wormhole 32; hal.get_arch_num_circular_buffers()), but the emulator capped
CB state at 32, so those kernels either failed to JIT-compile (a constexpr OOB
read of unpack_src_format[>=32] in reduce_helpers_compute.inl) or aborted at
runtime (cb_wait_front on a CB id >= 32 saw num_pages=0 from an out-of-bounds
cb_sync_states_ read).

Raise the CB count to 64 across the emulator:
- Core::MAX_CBS 32 -> 64 (device.hpp): sizes cb_sync_states_[] and un-gates
  init_cb_sync for ids >= 32 — the direct fix for the cb_wait_front abort.
- cb_api.h: size the JIT per-CB constexpr metadata arrays to EMULE_JIT_CB_SLOTS
  (64), filled directly from the runner's now-64-value defines (the runner emits
  EMULE_NUM_CBS values, defaulting unconfigured CBs to Invalid/32x32); widen the
  self-produce/consume CB masks to 1ull << cb_id (UB for cb_id >= 32 as uint32_t).
- emule_thread_ctx.h: cb_self_{produce,consume}_mask uint32_t -> uint64_t.
- asan_cb.h / asan_l1_checks.h: size the unconditionally-written ASAN CB
  bookkeeping arrays to 64 so cb_id >= 32 no longer writes out of bounds.

Pairs with the tt-metal runner change (EMULE_NUM_CBS 32->64 + matching ASAN
array definitions).

Validated on a 37-CB `cat_wide` case (36-source concat): runs on emule at
PCC 0.9999986, matching real Blackhole silicon (also 0.9999986) — a >32-CB
program that is numerically correct on both. softmax/rms_norm (<=32 CBs)
unaffected.

+++ include/jit_hw/internal/emule_thread_ctx.h
@@ -124,8 +124,9 @@ struct ThreadCommonCtx {
     uint32_t my_thread_id = 0;   // get_my_thread_id() (was __emule_my_thread_id)
 
     // Cross-role CB state (shared base — a CB producer/consumer may be either role).
-    uint32_t cb_self_consume_mask = 0;            // was __emule_cb_self_consume_mask
-    uint32_t cb_self_produce_mask = 0;            // was __emule_cb_self_produce_mask
+    // 64-bit so bit cb_id is valid for Blackhole's 64 CBs (1ull << cb_id, cb_id up to 63).
+    uint64_t cb_self_consume_mask = 0;            // was __emule_cb_self_consume_mask
+    uint64_t cb_self_produce_mask = 0;            // was __emule_cb_self_produce_mask
     LocalCBInterface local_cb[__EMULE_CTX_MAX_CBS]{};  // per-RISC CB ring ptrs (was __emule_local_cb)
 
     explicit ThreadCommonCtx(Kind k) : kind(k) {}
